requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.15
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.3
inflection==0.5.1
instaloader==4.9.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.1
openapi-codec==1.3.2
outcome==1.2.0
packaging==21.3
pyparsing==3.0.9
PySocks==1.7.1
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.5.0
simplejson==3.17.6
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.10.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
wsproto==1.2.0
